Promote cns argument to int64_t to ensure correctness
Some operations such as cns(-1) would end up emitting
the wrong thing, because the -1 literal would be interpreted as
int32_t and only four bytes would get copied across by memcpy.
The result is that cns(-1) would in fact emit the constant
4294967295 (= 2^32-1) which is bound to lead to wrong code.

A former workaround involved using cns(-1LL). I removed this
use case to show that this is now kosher (and added a test to
show that nothing broke - use TRACE=hhir:2 to check).