No longer needed to specify these defines on OSX.
This was a work-around for hhvm not finding jemalloc symbol.

The cause for this issue was an incorrect homebrew package. jemalloc was
being built with a prefix, and __*_hook's were never exported due to
lack of __GLIBC__ define.

A patched version has been provided on the following link:
https://gist.github.com/danslo/6034879/raw/00be8488d2dea232b750ab8207a788032d9135b6/jemallocfb.rb

It was needed to add the ability to specify jemalloc's include
directory, due to the homebrew package being keg_only.