Fix maybe-uninitialized errors and ext_image
GCC 4.8 picks up a number of maybe-uninitialized code. Most of
which has been properly fixed and one place that was supressed
using a pragma push and pop around the code.

This also updates the ext_image code to work with newer libraries,
since a number of functions are no longer exported from the
dynamic library.