Fix lock rank violation assersions
There's a race in the emitter, where if we emit a unit (because,
it wasn't in the repo) fail to commit it, and then find the unit in the repo,
we destroy the unit we just emitted, and read the one from the repo instead.

This typically happens at RankFileRepo. A new change to Func::~Func makes
it grab the write lease in DEBUG builds, which is a lock rank violation.

Newly created Funcs never need the write lease, however, so aquire it
lazily.

This was happening running slow tests, because some tests are run multiple
times (potentially in parallel) with different runtime options (in fact, the
test code is duplicated in separate files).