Set EnableXHP=0 by default, improve how the lexer works when EnableXHP=0
A couple of developers on github have reported that they've hit errors at
parse time with certain expressions that work under Zend PHP, or that they
have gotten output unexpected tokens from token_get_all(). A lot of this
is related to the lexer/grammar rules for XHP, so an easy solution to fix
a lot of these issues is to set EnableXHP=0 by default and to make some
changes to the lexer to make it more Zend PHP compatible when EnableXHP=0.

Closes #740
Closes #846

https://github.com/facebook/hiphop-php/issues/740
https://github.com/facebook/hiphop-php/issues/846

Differential Revision: D905986