(Re)move a few more unused/out-of-place functions on StringData
StringData::getPrecomputedHash was unused.  The
isValidVariableName thing was a thin wrapper on a zend function that
is hardly used, and it shouldn't be a member function anyway.  The
convert_double_helper and convert_integer_helper also shouldn't be
member functions, and the former duplicated buildStringData(double)
too much for me so I removed it.

Differential Revision: D903915