Tighten checkRegisters()
Currently, check registers works by traversing down the dominator
tree.  Effects on registers from a dominator block can be seen
in each dominator child.  This will miss a rogue assignment that
happens in a block that does not dominate one of its successors.
For example if we mis-assign to a register in one arm of an if-
then-else block, we won't notice.

This diff changes checkRegisters() to propagate register state
along flow edges, so we can catch such rogue assignments.

Differential Revision: D903949